[IA] Adding mouse move into map2

+++ Assets/Scripts/Character/Character.cs
@@ -3,7 +3,7 @@ using UnityEngine.UI;
 
 abstract public class Character : MonoBehaviour
 {
-    public int x, y, z;
+    public float x, y, z;
     public double width, height, depth;
     public float life, maxLife;
